+++ RELEASENOTES.md
@@ -14,6 +14,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Feature - Connect My App API on Home Page
 * Feature - App Release Process - Business logic & API connect for document upload
 * Feature - My App User Experience
+* Feature - App overviw - Show details overlay when Application status is in "in review"
 
 ### 0.5.5